% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Israel_mortality}
\alias{Israel_mortality}
\title{Monthly deaths in Israel between January 2001 and May 2021}
\format{
A data frame with 245 rows and 7 variables:
\describe{
  \item{Month}{The month number, between 1 to 12}
  \item{Year}{The year, between 2001 to 2021}
  \item{monthly_total}{Total monthly count of deaths for the entire population}
  \item{Date}{The date}
  \item{monthly_est}{Monthly estimated population size}
  \item{percent}{Monthly mortality percent for the entire population}
  \item{time}{A numeric vector counting the number of months (1 to 245)}
}
}
\source{
\url{https://www.cbs.gov.il}
}
\usage{
Israel_mortality
}
\description{
A dataset containing the monthly number of deaths, the estimated monthly population size, and the monthly mortality percent, for all the population and for groups of different age and sex.
}
\keyword{datasets}
