% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{its_lm}
\alias{its_lm}
\title{ITS analysis for continuous outcomes}
\usage{
its_lm(
  data,
  form,
  time_name,
  intervention_start_ind,
  freq,
  seasonality = "none",
  impact_model = "full",
  counterfactual = FALSE
)
}
\arguments{
\item{data}{The data frame corresponding to the supplied formula, existing of at least 2 variables: (1) the count outcome, and (2) a vector of time points}

\item{form}{A formula with the response on the left, followed by the ~ operator, and the covariates on the right, separated by + operators. The formula should not contain an offest term.}

\item{time_name}{A string giving the name of the time variable. The time variable may or may not be supplied as a covariate in the formula}

\item{intervention_start_ind}{Numeric - a number between 1 and nrow(data)-1 stating the time point of the start of the intervention}

\item{freq}{A positive integer describing the frequency of the time series.}

\item{seasonality}{A string specifying whether seasonality should be considered. Possible options include "none" corresponding to no seasonal adjustment, "full" corresponding to using freq-1 Fourier terms to model the seasonal component, and "significant" indicating whether only the significant Fourier terms should be considered in the seasonal adjustment. Default value is "none".}

\item{impact_model}{A string specifying the assumed impact model. Possible options include "full" corresponding to a model including both a level change and a slope change, "level" corresponding to a model including just a level change, and "slope" corresponding to a model including just a slope change. Default value is "full".}

\item{counterfactual}{Logical - indicating whether the model-based counterfactual values should also be returned as an additional column in the data. Default value is FALSE, in which case the counterfactual values are not returned.}
}
\value{
The function returns a list with three elements: the fitted linear regression model, the summary of the model (including the mean difference and Cohen's d), and the original data together with the model predictions.
}
\description{
\code{its_lm} fits a linear regression model to an ITS, and returns the model, the summary of the model (including the mean difference and Cohen's d), and the original data together with the model predictions.
}
\examples{
data <- unemployed
form <- as.formula("percent ~ time")
intervention_start_ind <- which(data$year==2020 & data$month>2| data$year==2021)[1]
fit <- its_lm(data=data,form=form,time_name = "time",intervention_start_ind=intervention_start_ind,freq=12, seasonality= "none", impact_model = "full",counterfactual = TRUE)
}
