% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{its_poisson_wo_seas}
\alias{its_poisson_wo_seas}
\title{ITS analysis for count outcomes with no seasonality adjustment}
\usage{
its_poisson_wo_seas(
  data,
  form,
  offset_name = NULL,
  time_name,
  intervention_start_ind,
  over_dispersion = FALSE,
  impact_model = "full",
  counterfactual = FALSE,
  print_summary = FALSE
)
}
\arguments{
\item{data}{The data frame corresponding to the supplied formula, existing of at least 2 variables: (1) the count outcome, and (2) a vector of time points}

\item{form}{A formula with the response on the left, followed by the ~ operator, and the covariates on the right, separated by + operators. The formula should not contain an offest term.}

\item{offset_name}{either a string indicating the name of the offset column in the data, or NULL. Default value is NULL}

\item{time_name}{A string giving the name of the time variable. The time variable may or may not be supplied as a covariate in the formula}

\item{intervention_start_ind}{Numeric - a number between 1 and nrow(data)-1 stating the time point of the start of the intervention}

\item{over_dispersion}{Logical - indicating whether a quasi-Poisson model should be used to account for over dispersion (TRUE), or not (FLASE). Default value is FALSE.}

\item{impact_model}{A string specifying the assumed impact model. Possible options include "full" corresponding to a model including both a level change and a slope change, "level" corresponding to a model including just a level change, and "slope" corresponding to a model including just a slope change. Default value is "full".}

\item{counterfactual}{Logical - indicating whether the model-based counterfactual values should also be returned as an additional column in the data. Default value is FALSE, in which case the counterfactual values are not returned.}

\item{print_summary}{Logical - indicating whethwe the entire model summary should be printed, or just the relevant effect size. Default value is FALSE in which case only the effect size, together with its 95\% CI and P-value, are printed.}
}
\value{
The function returns a list with three elements: the fitted Poisson regression model, the summary of the model (including the relative risk), and the original data together with the model predictions.
}
\description{
\code{its_poisson_wo_seas} fits a Poisson regression model to an ITS, and returns the model, the summary of the model (including the relative risk), and the original data together with the model predictions.
}
\examples{
data <- unemployed
form <- as.formula("unemployed ~ time")
intervention_start_ind <- which(data$year==2020 & data$month>2| data$year==2021)[1]
fit <- its_poisson_wo_seas(data=data,form=form,offset_name="labour", time_name = "time",intervention_start_ind=intervention_start_ind,over_dispersion=TRUE, impact_model = "full",counterfactual = TRUE)
}
