% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployed}
\alias{unemployed}
\title{Monthly unemployment count and percent in Israel between January 2013 and May 2021}
\format{
A data frame with 101 rows and 8 variables:
\describe{
  \item{year}{The year, between 2013 to 2021}
  \item{month}{The number month, between 1 to 12}
  \item{unemployed}{Total monthly count of unemployed}
  \item{labour}{Total monthly count of labour force}
  \item{percent}{Monthly unemployment percent}
  \item{dt}{The date}
  \item{time}{A numeric vector counting the number of months (1 to 101)}
  \item{indicator}{The Covid-19 binary indicator, equal to zero prior to March 2020, otherwise equal to 1}
}
}
\source{
\url{https://www.cbs.gov.il/en/publications/Pages/2021/Labour-Force-Survey-Monthly-May-2021.aspx}
}
\usage{
unemployed
}
\description{
A dataset containing the monthly unemployment count, the monthly labour force, and the monthly unemployment percent, as well as other attributes.
}
\keyword{datasets}
