% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zero_inflated_sim_data}
\alias{zero_inflated_sim_data}
\title{Simulated monthly count data with excess zeros between January 2001 and May 2021}
\format{
A data frame with 245 rows and 5 variables:
\describe{
  \item{Month}{The month number, between 1 to 12}
  \item{Year}{The year, between 2001 to 2021}
  \item{monthly_total}{Total simulated zero-inflated monthly count}
  \item{Date}{The date}
  \item{time}{A numeric vector counting the number of months (1 to 245)}
}
}
\usage{
zero_inflated_sim_data
}
\description{
A dataset containing the simulated monthly zero-inflated count.
}
\keyword{datasets}
